refactor:DogAnimationのアクションのアニメーションを分離
